[onert] Unify TensorBuilder impl of cpu backends
Unify TensorBuilder implementation of cpu-based backends.
They all use the same logic. This commit moves the common
implementation to `cpu_common` directory.

ONE-DCO-1.0-Signed-off-by: Hanjoung Lee <hanjoung.lee@samsung.com>
