+++ runtime/onert/backend/xnnpack/TensorBuilder.h
@@ -17,15 +17,7 @@
 #ifndef __ONERT_BACKEND_XNNPACK_TENSOR_BUILDER_H__
 #define __ONERT_BACKEND_XNNPACK_TENSOR_BUILDER_H__
 
-#include <backend/cpu_common/DynamicTensorManager.h>
-#include <backend/cpu_common/TensorRegistry.h>
-
-#include <ir/OperandIndexMap.h>
-
-#include "StaticTensorManager.h"
-#include "Tensor.h"
-
-#include <unordered_map>
+#include <backend/cpu_common/TensorBuilder.h>
 
 namespace onert
 {
@@ -34,39 +26,7 @@ namespace backend
 namespace xnnpack
 {
 
-class TensorBuilder
-{
-public:
-  TensorBuilder(const std::shared_ptr<cpu_common::TensorRegistry> &tensor_reg);
-
-  /**
-   * @brief     Register tensor information to allocate on XNNPACK backend
-   * @param[in] ind    Operand index
-   * @param[in] info   Operand information
-   * @param[in] layout Operand data layout
-   */
-  void registerTensorInfo(const ir::OperandIndex &ind, const ir::OperandInfo &info,
-                          ir::Layout backend_layout);
-
-  void notifyFirstUse(const ir::OperandIndex &);
-  void notifyLastUse(const ir::OperandIndex &);
-
-  bool isRegistered(const ir::OperandIndex &) const;
-
-  void prepare(void);
-  void allocate();
-  void postFunctionPrepare()
-  { /* DO NOTHING */
-  }
-
-  IDynamicTensorManager *dynamicTensorManager(void) { return _dynamic_tensor_mgr.get(); }
-
-private:
-  const std::shared_ptr<cpu_common::TensorRegistry> _tensor_reg;
-  std::unique_ptr<cpu_common::DynamicTensorManager> _dynamic_tensor_mgr;
-  std::unique_ptr<StaticTensorManager> _static_tensor_mgr;
-  ir::OperandIndexMap<ir::OperandInfo> _tensor_info_map;
-};
+using TensorBuilder = cpu_common::TensorBuilder;
 
 } // namespace xnnpack
 } // namespace backend
